fix pagination info in server mode
I changed hypen to endash for the pagination display. In server mode, the last index of items was showing the amount for a full page, but often it's less than a full page, eg 12 items when page size is ten, you show 10-15 of 12, should be 10-12 of 12. Maybe you didn't notice this because your doco page lists 500 items, if you list 497 instead you will see the problem.